feat: Introduce AbstractAviatorSSCSastAuditCommand class for SAST audit functionality
